moved watchdog functions into separate files. editted main.cpp to reflect these changes.
