+++ fsae-vehicle-fw/src/peripherals/watchdog.h
@@ -1,0 +1,8 @@
+#ifndef WATCHDOG_H
+#define WATCHDOG_H
+
+void initWatchdog();
+
+void Watchdog_Pet();
+
+#endif WATCHDOG_H
